Change declared version of Java for the project: 1.5 -> 1.6
The code already uses a number of methods that are unavailable in Java 5, were added in Java 6.
As such, it cannot be compiled with Java 5 and cannot run on Java 5.

I doubt that anybody is interested in Java 1.5 nowadays, thus updating pom.xml instead of fixing the code.

The methods:
- constructor java.io.IOException(String, Throwable)
- java.util.Arrays.copyOfRange(int[], int, int)
- java.util.Arrays.copyOfRange(T[], int, int), used with <T> = String
- java.lang.String.isEmpty()